TDB-129 : Normalize TokuDB ALTER code to remove 5.5 and version dependent compilation
- Moved content of ha_tokudb_alter_common.cc into ha_tokudb_alter_56.cc
  and renamed ha_tokudb_alter_56.cc to ha_tokudb_alter.cc. Reformatted moved
  code and cleaned up numerous whitespace violations.
- Removed all ALTER_55 and ALTER_56 conditional compilation to leave inly the
  ALTER_56 code behind.
- Removed some MARIADB compilation specific code in alter pathways.
